requirements.txt
Django~=4.2.10
asgiref~=3.7.2
pip~=24.0
setuptools~=69.1.1
sqlparse~=0.4.4
typing-extensions~=4.9.0
tzdata~=2024.1
beautifulsoup4~=4.12.3
django-bootstrap4~=24.1
whitenoise~=6.6.0
Pillow~=10.2.0
selenium~=4.18.1
cloudinary~=1.38.0
django-cloudinary-storage~=0.3.0
python-decouple~=3.8
urllib3>=1.25,<2.0
psycopg2-binary~=2.9.9